requirements.txt
streamlit==1.37.1
google-generativeai==0.7.2
pypdf==4.3.1
python-docx==1.1.2
Pillow==10.4.0
plotly==5.23.0
pandas==2.2.2
graphviz==0.20.3
